Probe pkg stapling before notarizing on runs>0
